vr4300: Sign extend results from MFC0.
This bug prevented Conker's Bad Fur Day from booting.

+++ vr4300/cp0.c
@@ -136,12 +136,12 @@ int VR4300_MFC0(struct vr4300 *vr4300,
   unsigned src = GET_RD(iw);
 
   if (src == (VR4300_CP0_REGISTER_COUNT - 32)) {
-    exdc_latch->result = (uint32_t)
+    exdc_latch->result = (int32_t)
       (vr4300->regs[VR4300_CP0_REGISTER_COUNT] >> 1);
   }
 
   else
-    exdc_latch->result = mask_reg(src, vr4300->regs[32 + src]);
+    exdc_latch->result = (int32_t) mask_reg(src, vr4300->regs[32 + src]);
 
   exdc_latch->dest = (int32_t) dest;
   return 0;
